refactor: outsource prevent theme flash component
